requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
gunicorn==23.0.0
idna==3.10
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.1.0
requests==2.32.3
scipy==1.15.2
sqlparse==0.5.3
urllib3==2.4.0
